+++ Space.DTO/Spatial/SpatialEntity.cs
@@ -22,9 +22,9 @@ namespace Space.DTO.Spatial
 
         #region ISpatialEntity
 
-        public float Latitude { get; set; }
+        public double Latitude { get; set; }
 
-        public float Longitude { get; set; }
+        public double Longitude { get; set; }
 
         #endregion ISpatialEntity
 
@@ -33,23 +33,23 @@ namespace Space.DTO.Spatial
         /// <summary>
         /// The mass of the spatial entity
         /// </summary>
-        public float Mass { get; set; }
+        public double Mass { get; set; }
 
         /// <summary>
         /// The radius of the entity.  Used for scaling the planet's image
         /// </summary>
-        public float Radius { get; set; }
+        public double Radius { get; set; }
 
         /// <summary>
         /// The distance for the obit from the most massive entity.
         /// </summary>
-        public float OrbitRadius { get; set; }
+        public double OrbitRadius { get; set; }
 
         /// <summary>
         /// The speed as the planet orbits around the most massive entity.
         /// We could do a fancy calculation here but it's much nicer to do it this way
         /// </summary>
-        public float OrbitSpeed { get; set; }
+        public double OrbitSpeed { get; set; }
 
         #endregion
     }
