+++ Space.DTO/Units/Agent.cs
@@ -13,8 +13,8 @@ namespace Space.DTO.Units
 
         #region Implementation of ISpatialEntity
 
-        public float Latitude { get; set; }
-        public float Longitude { get; set; }
+        public double Latitude { get; set; }
+        public double Longitude { get; set; }
 
         public int FleetID { get; set; }
