+++ Space.Repository.Entities/ISpatialEntity.cs
@@ -5,11 +5,11 @@
         /// <summary>
         /// The latitude of the entity in space
         /// </summary>
-        float Latitude { get; set; }
+        double Latitude { get; set; }
 
         /// <summary>
         /// The longitude of the entity in space
         /// </summary>
-        float Longitude { get; set; }
+        double Longitude { get; set; }
     }
 }
